# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlationId": {
            "title": "Correlationid",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "AuditResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "events"
        ],
        "title": "AuditResponse",
        "type": "object"
      },
...
      "DemoCatalog": {
        "additionalProperties": false,
        "properties": {
          "actors": {
            "items": {
              "$ref": "#/components/schemas/Actor"
            },
            "title": "Actors",
            "type": "array"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/DemoCase"
            },
            "title": "Cases",
            "type": "array"
          },
          "defaultCaseId": {
            "title": "Defaultcaseid",
            "type": "string"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          }
        },
        "required": [
          "cases",
          "actors",
          "disclaimer",
          "defaultCaseId"
        ],
        "title": "DemoCatalog",
        "type": "object"
      },
      "DemoLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Actorid",
            "type": "string"
          }
        },
        "required": [
          "actorId"
        ],
        "title": "DemoLoginRequest",
        "type": "object"
      },
      "DemoLoginResponse": {
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "tokenType": {
            "const": "bearer",
            "title": "Tokentype",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "actor",
          "expiresAt"
        ],
        "title": "DemoLoginResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NurseActionRequest": {
        "additionalProperties": false,
        "properties": {
          "actionId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Actionid",
            "type": "string"
          },
          "actionType": {
            "$ref": "#/components/schemas/ActionType"
          },
          "baseStateVersion": {
            "minimum": 0.0,
            "title": "Basestateversion",
            "type": "integer"
          },
          "overrideDisposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Disposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "rationale": {
            "maxLength": 3000,
            "minLength": 3,
            "title": "Rationale",
            "type": "string"
          },
          "statedPreference": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statedpreference"
          },
          "targetRecommendationId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Targetrecommendationid",
            "type": "string"
          }
        },
        "required": [
          "actionId",
          "targetRecommendationId",
          "baseStateVersion",
          "actionType",
          "rationale"
        ],
        "title": "NurseActionRequest",
        "type": "object"
      },
...
      "SessionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "caseId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Caseid",
            "type": "string"
          }
        },
        "required": [
          "caseId"
        ],
        "title": "SessionCreateRequest",
        "type": "object"
      },
...
      "SessionSummary": {
        "additionalProperties": false,
        "properties": {
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "patientName": {
            "title": "Patientname",
            "type": "string"
          },
          "readinessState": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "stateVersion": {
            "title": "Stateversion",
            "type": "integer"
          }
        },
        "required": [
          "sessionId",
          "patientName",
          "caseId",
          "stateVersion",
          "readinessState"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/NurseAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "careTeamId": {
            "title": "Careteamid",
            "type": "string"
          },
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "latestResponse": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "nurseNotes": {
            "title": "Nursenotes",
            "type": "string"
          },
          "originalTranscript": {
            "title": "Originaltranscript",
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientContext"
          },
          "patientOutcome": {
            "enum": [
              "UNRECORDED",
              "CARE_ACCEPTED",
              "REMAINS_HOME",
              "CLINICIAN_REVIEW"
            ],
            "title": "Patientoutcome",
            "type": "string"
          },
          "patientResponses": {
            "items": {
              "$ref": "#/components/schemas/PatientResponse"
            },
            "title": "Patientresponses",
            "type": "array"
          },
          "permissions": {
            "$ref": "#/components/schemas/SessionPermissions"
          },
          "recommendationStatus": {
            "enum": [
              "AWAITING_REVIEW",
              "ACCEPTED",
              "DECLINED_BY_PATIENT",
              "ESCALATED",
              "OVERRIDDEN"
            ],
            "title": "Recommendationstatus",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/SystemRecommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TriageState"
          }
        },
        "required": [
          "sessionId",
          "caseId",
          "careTeamId",
          "patient",
          "originalTranscript",
          "nurseNotes",
          "state",
          "latestResponse",
          "recommendations",
          "actions",
          "patientResponses",
          "recommendationStatus",
          "patientOutcome",
          "permissions"
        ],
        "title": "SessionView",
        "type": "object"
      },
...
      "TriageRequest": {
        "additionalProperties": false,
        "properties": {
          "baseStateVersion": {
            "minimum": 0.0,
            "title": "Basestateversion",
            "type": "integer"
          },
          "nurseNotes": {
            "anyOf": [
              {
                "maxLength": 6000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nursenotes"
          },
          "patientDemographics": {
            "$ref": "#/components/schemas/PatientDemographics"
          },
          "patientId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Patientid",
            "type": "string"
          },
          "sessionId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Sessionid",
            "type": "string"
          },
          "symptomDescription": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Symptomdescription",
            "type": "string"
          },
          "turnId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Turnid",
            "type": "string"
          },
          "vitalSigns": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalSigns"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "sessionId",
          "turnId",
          "baseStateVersion",
          "patientId",
          "symptomDescription",
          "patientDemographics"
        ],
        "title": "TriageRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Synthetic demonstration only. Not a medical device or real clinical guidance.",
    "title": "Meridian Patient Navigation & Triage Copilot",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/demo": {
      "post": {
        "operationId": "demoLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Demo Login"
      }
    },
    "/api/demo": {
      "get": {
        "operationId": "getDemoCatalog",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Demo Catalog"
      }
    },
...
    "/api/sessions": {
      "get": {
        "operationId": "listSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Listsessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions"
      },
      "post": {
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Session"
      }
    },
    "/api/sessions/{sessionId}": {
      "get": {
        "operationId": "getSession",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session"
      }
    },
    "/api/sessions/{sessionId}/actions": {
      "post": {
        "operationId": "recordNurseAction",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NurseActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Action"
      }
    },
    "/api/sessions/{sessionId}/audit": {
      "get": {
        "operationId": "getAudit",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit"
      }
    },
    "/api/triage": {
      "post": {
        "operationId": "runTriage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Triage"
      }
    }
  }
}